Landmark Recognition (CSC420)/requirements.txt
opencv_contrib_python==3.4.2.17
torch==1.2.0
numpy==1.17.2
pandas==0.25.1
matplotlib==3.1.1
torchvision==0.4.0a0
Pillow==6.2.1
